refactor: extract components from overview, help pages
- overview/kpi-card.tsx, chart-tooltips.tsx
- help/faq-item.tsx, help-primitives.tsx (CodeBlock, HelpSectionTitle)
