+++ apps/web/src/components/help/index.ts
@@ -1,0 +1,2 @@
+export { FaqItem } from './faq-item';
+export { CodeBlock, HelpSectionTitle } from './help-primitives';
