+++ apps/web/src/pages/dashboard/overview.tsx
@@ -52,118 +52,9 @@ import { useThemeStore } from '../../store/theme.store';
 import { cn } from '../../lib/utils';
 import { useTranslation } from 'react-i18next';
 import { toast } from 'sonner';
+import { KpiCard, PnlTooltip, AssetTooltip } from '../../components/overview';
 
 gsap.registerPlugin(useGSAP);
-
-// ── KPI Card ──────────────────────────────────────────────────────────────────
-function KpiCard({
-  label,
-  value,
-  sub,
-  positive,
-  icon,
-  tooltip,
-  className,
-  href,
-}: {
-  label: string;
-  value: string | number;
-  sub?: string;
-  positive?: boolean;
-  icon: React.ReactNode;
-  tooltip?: string;
-  className?: string;
-  href?: string;
-}) {
-  const inner = (
-    <div
-      className={cn(
-        'stat-card h-full rounded-xl border border-border bg-card p-5',
-        href &&
-          'cursor-pointer transition-colors hover:border-primary/40 hover:bg-card/80',
-        className,
-      )}
-    >
-      <div className="mb-3 flex items-center justify-between">
-        <div className="flex items-center gap-1.5 text-sm text-muted-foreground">
-          {label}
-          {tooltip && <InfoTooltip text={tooltip} />}
-        </div>
-        <div className="rounded-lg bg-muted p-2 text-muted-foreground">
-          {icon}
-        </div>
-      </div>
-      <div
-        className={cn(
-          'text-2xl font-bold',
-          positive === true
-            ? 'text-emerald-500'
-            : positive === false
-              ? 'text-red-500'
-              : '',
-        )}
-      >
-        {value}
-      </div>
-      {sub && <div className="mt-1 text-xs text-muted-foreground">{sub}</div>}
-    </div>
-  );
-  return href ? (
-    <Link to={href} className="block h-full">
-      {inner}
-    </Link>
-  ) : (
-    inner
-  );
-}
-
-// ── PnL Tooltip ───────────────────────────────────────────────────────────────
-function PnlTooltip({
-  active,
-  payload,
-  label,
-}: Partial<TooltipContentProps<number, string>>) {
-  if (!active || !payload?.length) return null;
-  const val = payload[0].value as number;
-  const isPos = val >= 0;
-  return (
-    <div className="rounded-xl border border-border bg-card px-3 py-2 shadow-xl text-xs">
-      <p className="text-muted-foreground mb-1">{label}</p>
-      <p
-        className={cn(
-          'font-bold text-sm',
-          isPos ? 'text-emerald-400' : 'text-red-400',
-        )}
-      >
-        {isPos ? '+' : ''}${val.toFixed(2)}
-      </p>
-    </div>
-  );
-}
-
-// ── Asset Tooltip ─────────────────────────────────────────────────────────────
-function AssetTooltip({
-  active,
-  payload,
-  label,
-}: Partial<TooltipContentProps<number, string>>) {
-  if (!active || !payload?.length) return null;
-  const val = payload[0].value as number;
-  const isPos = val >= 0;
-  return (
-    <div className="rounded-xl border border-border bg-card px-3 py-2 shadow-xl text-xs">
-      <p className="font-semibold mb-1">{label}</p>
-      <p
-        className={cn(
-          'font-bold text-sm',
-          isPos ? 'text-emerald-400' : 'text-red-400',
-        )}
-      >
-        {isPos ? '+' : ''}${val.toFixed(2)} P&L
-      </p>
-    </div>
-  );
-}
 
 // ── Page ──────────────────────────────────────────────────────────────────────
 export function OverviewPage() {
