+++ apps/web/src/pages/help.tsx
@@ -3,9 +3,7 @@ import {
   HelpCircle,
   BookOpen,
   Key,
-  ChevronDown,
   ChevronRight,
-  ChevronUp,
   ExternalLink,
   AlertTriangle,
   GitFork,
@@ -28,76 +26,7 @@ import {
   ExplainPanel,
 } from '@crypto-trader/ui';
 import { AgentsShowcaseSection } from './dashboard/agents-showcase';
-
-function FaqItem({ q, a }: { q: string; a: string }) {
-  const [open, setOpen] = useState(false);
-  const bodyRef = useRef<HTMLDivElement>(null);
-
-  useGSAP(
-    () => {
-      if (!bodyRef.current) return;
-      if (open) {
-        gsap.fromTo(
-          bodyRef.current,
-          { height: 0, opacity: 0 },
-          { height: 'auto', opacity: 1, duration: 0.3, ease: 'power2.out' },
-        );
-      } else {
-        gsap.to(bodyRef.current, {
-          height: 0,
-          opacity: 0,
-          duration: 0.2,
-          ease: 'power2.in',
-        });
-      }
-    },
-    { dependencies: [open] },
-  );
-
-  return (
-    <div className="rounded-lg border border-border overflow-hidden">
-      <button
-        onClick={() => setOpen((v) => !v)}
-        className="flex w-full items-center justify-between gap-3 px-4 py-3.5 text-left font-medium text-sm hover:bg-muted/30 transition-colors"
-      >
-        {q}
-        {open ? (
-          <ChevronUp className="h-4 w-4 shrink-0 text-primary" />
-        ) : (
-          <ChevronDown className="h-4 w-4 shrink-0 text-muted-foreground" />
-        )}
-      </button>
-      <div ref={bodyRef} className="overflow-hidden h-0 opacity-0">
-        <p className="px-4 pb-4 text-sm text-muted-foreground leading-relaxed">
-          {a}
-        </p>
-      </div>
-    </div>
-  );
-}
-
-function CodeBlock({ children }: { children: React.ReactNode }) {
-  return (
-    <code className="block rounded-lg bg-muted px-4 py-3 font-mono text-xs text-foreground">
-      {children}
-    </code>
-  );
-}
-
-function SectionTitle({
-  icon,
-  children,
-}: {
-  icon: React.ReactNode;
-  children: React.ReactNode;
-}) {
-  return (
-    <div className="mb-5 flex items-center gap-2 border-b border-border pb-3">
-      <span className="text-primary">{icon}</span>
-      <h2 className="text-xl font-bold">{children}</h2>
-    </div>
-  );
-}
+import { FaqItem, CodeBlock, HelpSectionTitle as SectionTitle } from '../components/help';
 
 const SECTION_IDS = [
   'faq',
